Rename modules folder to views
